requirements.txt
colorama==0.4.6
cryptg==0.5.0.post0
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.1.1
rsa==4.9.1
Telethon==1.40.0
tqdm==4.67.1
